+++ tango/integrations/wandb/torch_train_callback.py
@@ -123,7 +123,7 @@ class WandbTrainCallback(TrainCallback):
             self.wandb.finish()
 
     @overrides
-    def post_batch(self, step: int, batch_loss: float) -> None:
+    def log_batch(self, step: int, batch_loss: float) -> None:
         if self.is_local_main_process:
             self.wandb.log(
                 {"train/loss": batch_loss, "train/lr": self.optimizer.param_groups[0]["lr"]},
@@ -131,6 +131,12 @@ class WandbTrainCallback(TrainCallback):
             )
 
     @overrides
-    def post_val_loop(self, step: int, val_metric_name: str, val_metric: float) -> None:
+    def post_val_loop(self, step: int, val_metric: float, best_val_metric: float) -> None:
         if self.is_local_main_process:
-            self.wandb.log({f"val/{val_metric_name}": val_metric}, step=step)
+            self.wandb.log(
+                {
+                    f"val/{self.train_config.val_metric_name}": val_metric,
+                    f"val/best_{self.train_config.val_metric_name}": best_val_metric,
+                },
+                step=step,
+            )