+++ CHANGELOG.md
@@ -7,6 +7,45 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.3.6](https://github.com/allenai/tango/releases/tag/v0.3.6) - 2021-11-12
+
+### Added
+
+- Added a `.log_batch()` method on `torch::TrainCallback` which is given the average loss across
+  distributed workers, but only called every `log_every` steps.
+
+### Removed
+
+- Removed `.pre_log_batch()` method on `torch::TrainCallback`.
+
+### Fixed
+
+- Fixed typo in parameter name `remove_stale_checkpoints` in `TorchTrainStep` (previously was `remove_state_checkpoints`).
+- Fixed bug in `FromParams` that would cause failures when `from __future__ import annotations`
+  was used with Python older than 3.10. See [PEP 563](https://www.python.org/dev/peps/pep-0563/)
+  for details.
+
+## [v0.3.5](https://github.com/allenai/tango/releases/tag/v0.3.5) - 2021-11-05
+
+### Fixed
+
+- Fixed a bug in `FromParams` where the "type" parameter was ignored in some cases
+  where the `Registrable` base class did not directly inherit from `Registrable`.
+
+## [v0.3.4](https://github.com/allenai/tango/releases/tag/v0.3.4) - 2021-11-04
+
+### Added
+
+- Added `StopEarlyCallback`, a `torch::TrainCallback` for early stopping.
+- Added parameter `remove_stale_checkpoints` to `TorchTrainStep`.
+
+### Changed
+
+- Minor changes to `torch::TrainCallback` interface.
+- Weights & Biases `torch::TrainCallback` now logs best validation metric score.
+
+## [v0.3.3](https://github.com/allenai/tango/releases/tag/v0.3.3) - 2021-11-04
+
 ### Added
 
 - Added support for PEP 604 in `FromParams`, i.e. writing union types as "X | Y" instead of "Union[X, Y]".
@@ -17,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a bug in `FromParams` where non-`FromParams` class parameters were not instantiated
   properly (or at all).
+- Fixed a bug in `FromParams` where kwargs were not passed on from a wrapper class to the wrapped class.
 - Fixed small bug where some errors from git would be printed when executor metadata is created
   outside of a git repository.